chore (Server): Move CRUD logic to dedicated controller
